refactor : TableService 에서 OrderRepository 참조 제거
