+++ src/main/java/kitchenpos/application/menu/MenuService.java
@@ -1,21 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.application.menu;
 
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import kitchenpos.domain.menu.Menu;
-import kitchenpos.domain.menu.MenuMapper;
-import kitchenpos.domain.menugroup.MenuGroup;
-import kitchenpos.domain.menugroup.MenuGroupRepository;
 import kitchenpos.domain.menu.MenuRepository;
-import kitchenpos.domain.product.Product;
-import kitchenpos.domain.product.ProductRepository;
-import kitchenpos.dto.MenuProductRequest;
 import kitchenpos.dto.MenuRequest;
 
 @Service
