+++ src/main/resources/application-test.properties
@@ -1,4 +1,3 @@
 spring.flyway.enabled=false
 spring.datasource.url=jdbc:h2:mem:testdb;MODE=MySQL
 spring.datasource.driverClassName=org.h2.Driver
-spring.jpa.hibernate.ddl-auto=validate
